HITL/requirements.txt
streamlit>=1.31.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
openai>=1.0.0